backend/services/notification-service/requirements.txt

pika
python-dotenv

gunicorn
Flask==2.3.3
python-dotenv==1.0.1
sendgrid==6.11.0